fix: ゲーム対戦画面の全てのヘッダーとフッターを削除
ゲームプレイに集中できるよう、対戦画面からヘッダーとフッターを完全に削除しました。

前回の修正で、一部のゲーム画面にヘッダーが残ってしまう問題がありましたが、本修正で`GameLayout`コンポーネント内のヘッダーも削除し、すべてのゲーム画面でヘッダーとフッターが非表示になるようにしました。

- `AppLayout` コンポーネントを新設し、URLのパスに応じてサイト全体のヘッダーとフッターの表示を切り替えるようにしました。
- `GameLayout` コンポーネントから、ゲーム固有のヘッダー表示を削除しました。
- これにより、ゲームページ (`/games/*`) では全てのヘッダーとフッターが非表示になります。
- トップページやその他のページでは、これまで通りヘッダーとフッターが表示されます。
